Removed RAM queue, disk queue only for now.
The idea before this change was to have dirty (not committed)
logs in RAM but keep only committed logs on disk so that we
don't have to truncate already committed entries.
This turn out to be a mistake as there is no way to guarantee
that. e.g. when a leader commits some entries on disk then
crashes before replicating them to followers,
these logs will have to be truncated/overwritten when it comes
up again as follower.